Reorganization and Bug Fixes
Moved around specific css & js files to be where the html file is for better organization. Changed side navigation pannel to be loaded from the javascript instead of from every page so that less text needs to be edited in order to change every page.

Fixed a couple bugs in the armor and starting equipment tools, including movement speed not displaying correctly and copper pieces resulting in errors.

Added a quantity feature to the starting equipment to make it easier to deal with duplicate items.